requirements.txt
google-resumable-media
google-cloud-storage
google-cloud-bigquery
pandas
numpy==1.22.3
scikit-learn==1.1.1